+++ src/orquestra/sdk/_base/_driver/_ce_runtime.py
@@ -24,6 +24,25 @@ from orquestra.sdk.schema.workflow_run import (
 from . import _client, _exceptions, _models
 
 
+def _get_max_resources(workflow_def: WorkflowDef) -> _models.Resources:
+    max_gpu = None
+    max_memory = None
+    max_cpu = None
+    for inv in workflow_def.task_invocations.values():
+        if inv.resources is None:
+            continue
+        if inv.resources.memory is not None:
+            if parse_quantity(inv.resources.memory) > parse_quantity(max_memory or "0"):
+                max_memory = inv.resources.memory
+        if inv.resources.cpu is not None:
+            if parse_quantity(inv.resources.cpu) > parse_quantity(max_cpu or "0"):
+                max_cpu = inv.resources.cpu
+        if inv.resources.gpu is not None:
+            if int(inv.resources.gpu) > int(max_gpu or "0"):
+                max_gpu = inv.resources.gpu
+    return _models.Resources(cpu=max_cpu, memory=max_memory, gpu=max_gpu, nodes=None)
+
+
 class CERuntime(RuntimeInterface):
     """
     A runtime for communicating with the Compute Engine API endpoints
@@ -88,28 +107,20 @@ class CERuntime(RuntimeInterface):
             the workflow run ID
         """
 
-        max_gpu = None
-        max_memory = None
-        max_cpu = None
-        for inv in workflow_def.task_invocations.values():
-            if inv.resources is None:
-                continue
-            if inv.resources.memory is not None:
-                if parse_quantity(inv.resources.memory) > parse_quantity(
-                    max_memory or "0"
-                ):
-                    max_memory = inv.resources.memory
-            if inv.resources.cpu is not None:
-                if parse_quantity(inv.resources.cpu) > parse_quantity(max_cpu or "0"):
-                    max_cpu = inv.resources.cpu
-            if inv.resources.gpu is not None:
-                if int(inv.resources.gpu) > int(max_gpu or "0"):
-                    max_gpu = inv.resources.gpu
+        if workflow_def.resources is not None:
+            resources = _models.Resources(
+                cpu=workflow_def.resources.cpu,
+                memory=workflow_def.resources.memory,
+                gpu=workflow_def.resources.gpu,
+                nodes=workflow_def.resources.nodes,
+            )
+        else:
+            resources = _get_max_resources(workflow_def)
+
         try:
             workflow_def_id = self._client.create_workflow_def(workflow_def)
             workflow_run_id = self._client.create_workflow_run(
-                workflow_def_id,
-                _models.Resources(cpu=max_cpu, memory=max_memory, gpu=max_gpu),
+                workflow_def_id, resources
             )
         except _exceptions.InvalidWorkflowDef as e:
             raise exceptions.WorkflowSyntaxError(