+++ tests/sdk/v2/test_artifact_future_methods.py
@@ -12,21 +12,47 @@ from orquestra.sdk import exceptions
 from orquestra.sdk._base import _workflow
 from orquestra.sdk._base._dsl import DEFAULT_IMAGE, ArtifactFuture
 
-resources_no_default = {
+_TaskResourcesArgs = t.TypedDict(
+    "_TaskResourcesArgs",
+    {
+        "cpu": t.Optional[str],
+        "memory": t.Optional[str],
+        "disk": t.Optional[str],
+        "gpu": t.Optional[str],
+    },
+    total=False,
+)
+
+
+TASK_RESOURCES_ARGS: _TaskResourcesArgs = {
     "cpu": "2000m",
     "memory": "100Gi",
     "disk": "99Gi",
     "gpu": "1",
 }
 
-custom_image_no_default = {
+TASK_RESOURCES_EXPECTED = {
+    "cpu": "2000m",
+    "memory": "100Gi",
+    "disk": "99Gi",
+    "gpu": "1",
+    "nodes": None,
+}
+
+CUSTOM_IMAGE_NOT_DEFAULT = {
     "custom_image": "zapatacomputing/orquestra-not-default",
 }
 
-metadata_no_default = {**resources_no_default, **custom_image_no_default}
+CUSTOM_METADATA = {
+    "cpu": "2000m",
+    "memory": "100Gi",
+    "disk": "99Gi",
+    "gpu": "1",
+    "custom_image": "zapatacomputing/orquestra-not-default",
+}
 
 
-@sdk.task(resources=sdk.Resources(**resources_no_default))
+@sdk.task(resources=sdk.Resources(**TASK_RESOURCES_ARGS))
 def _resourced_task(ii: int = 0) -> int:
     return ii
 
@@ -40,7 +66,7 @@ class ObjWithTask:
     def __init__(self, ii: int = 0) -> None:
         self._ii = ii
 
-    @sdk.task(resources=sdk.Resources(**resources_no_default))
+    @sdk.task(resources=sdk.Resources(**TASK_RESOURCES_ARGS))
     def get_id(
         self,
     ) -> int:
@@ -48,7 +74,7 @@ class ObjWithTask:
 
 
 def get_task(ind: bool = True):
-    @sdk.task(resources=sdk.Resources(**resources_no_default))
+    @sdk.task(resources=sdk.Resources(**TASK_RESOURCES_ARGS))
     def task_get_number(number: int = 0):
         return number
 
@@ -63,28 +89,30 @@ def two_ouputs_task():
 @sdk.workflow
 def wf_assign_resources_to_deconstructed_task_invocation():
     result1, result2 = two_ouputs_task()
-    result1.with_resources(cpu=resources_no_default["cpu"])
+    result1.with_resources(cpu=TASK_RESOURCES_ARGS["cpu"])
     return [result1, result2]
 
 
 @sdk.workflow
 def wf_deconstruc_task_invocation_with_resources_assignment():
     result = two_ouputs_task()
-    result1, result2 = result.with_resources(cpu=resources_no_default["cpu"])
+    result1, result2 = result.with_resources(cpu=TASK_RESOURCES_ARGS["cpu"])
     return [result1, result2]
 
 
 @sdk.workflow
 def wf_assign_custom_image_to_deconstructed_task_invocation():
     result1, result2 = two_ouputs_task()
-    result1.with_custom_image(custom_image_no_default["custom_image"])
+    result1.with_custom_image(CUSTOM_IMAGE_NOT_DEFAULT["custom_image"])
     return [result1, result2]
 
 
 @sdk.workflow
 def wf_deconstruc_task_invocation_with_custom_image_assignment():
     result = two_ouputs_task()
-    result1, result2 = result.with_custom_image(custom_image_no_default["custom_image"])
+    result1, result2 = result.with_custom_image(
+        CUSTOM_IMAGE_NOT_DEFAULT["custom_image"]
+    )
     return [result1, result2]
 
 
@@ -92,8 +120,8 @@ def wf_deconstruc_task_invocation_with_custom_image_assignment():
 def wf_assign_metadata_to_deconstructed_task_invocation():
     result1, result2 = two_ouputs_task()
     result1.with_invocation_meta(
-        custom_image=custom_image_no_default["custom_image"],
-        gpu=resources_no_default["gpu"],
+        custom_image=CUSTOM_IMAGE_NOT_DEFAULT["custom_image"],
+        gpu=TASK_RESOURCES_ARGS["gpu"],
     )
     return [result1, result2]
 
@@ -102,8 +130,8 @@ def wf_assign_metadata_to_deconstructed_task_invocation():
 def wf_deconstruc_task_invocation_with_metadata_assignment():
     result = two_ouputs_task()
     result1, result2 = result.with_invocation_meta(
-        custom_image=custom_image_no_default["custom_image"],
-        gpu=resources_no_default["gpu"],
+        custom_image=CUSTOM_IMAGE_NOT_DEFAULT["custom_image"],
+        gpu=TASK_RESOURCES_ARGS["gpu"],
     )
     return [result1, result2]
 
@@ -117,7 +145,7 @@ def wf_task_call():
 def wf_with_resources_call():
     future = _task_without_resources()
     assert isinstance(future, ArtifactFuture)
-    new_future = future.with_resources(**resources_no_default)
+    new_future = future.with_resources(**TASK_RESOURCES_ARGS)
     return [future, new_future]
 
 
@@ -125,7 +153,7 @@ def wf_with_resources_call():
 def wf_with_custom_image_call():
     future = _task_without_resources()
     assert isinstance(future, ArtifactFuture)
-    new_future = future.with_custom_image(custom_image_no_default["custom_image"])
+    new_future = future.with_custom_image(CUSTOM_IMAGE_NOT_DEFAULT["custom_image"])
     return [future, new_future]
 
 
@@ -133,7 +161,7 @@ def wf_with_custom_image_call():
 def wf_with_invocation_meta_call():
     future = _task_without_resources()
     assert isinstance(future, ArtifactFuture)
-    new_future = future.with_invocation_meta(**metadata_no_default)
+    new_future = future.with_invocation_meta(**CUSTOM_METADATA)
     return [future, new_future]
 
 
@@ -141,8 +169,8 @@ def wf_with_invocation_meta_call():
 def wf_with_resources_and_custom_image_call():
     future = _task_without_resources()
     assert isinstance(future, ArtifactFuture)
-    new_future = future.with_resources(**resources_no_default).with_custom_image(
-        **custom_image_no_default
+    new_future = future.with_resources(**TASK_RESOURCES_ARGS).with_custom_image(
+        **CUSTOM_IMAGE_NOT_DEFAULT
     )
     return [future, new_future]
 
@@ -160,7 +188,7 @@ def wf_with_method_task():
 def wf_with_mixed_calls():
     simple_obj = ObjWithTask(13)
     simple_obj.get_id()
-    return _task_without_resources().with_invocation_meta(**metadata_no_default)
+    return _task_without_resources().with_invocation_meta(**CUSTOM_METADATA)
 
 
 class TestArifactFutureMethodsCalls:
@@ -290,8 +318,8 @@ class TestArifactFutureMethodsCalls:
         assert len(set(invocation.task_id for invocation in invocations)) == 1
 
         assert invocations[0].resources is not None
-        assert invocations[0].resources.dict() == resources_no_default
-        assert invocations[0].custom_image == custom_image_no_default["custom_image"]
+        assert invocations[0].resources.dict() == TASK_RESOURCES_EXPECTED
+        assert invocations[0].custom_image == CUSTOM_IMAGE_NOT_DEFAULT["custom_image"]
 
         assert invocations[1].resources is None
         assert invocations[1].custom_image is DEFAULT_IMAGE
@@ -310,7 +338,7 @@ class TestArifactFutureMethodsCalls:
         invocations = [*wf_model.task_invocations.values()]
         assert len(set(invocation.task_id for invocation in invocations)) == 1
         assert invocations[0].resources is not None
-        assert invocations[0].resources.dict() == resources_no_default
+        assert invocations[0].resources.dict() == TASK_RESOURCES_EXPECTED
 
         assert invocations[1].resources is None
         assert invocations[1].custom_image == invocations[0].custom_image
@@ -328,7 +356,7 @@ class TestArifactFutureMethodsCalls:
 
         invocations = [*wf_model.task_invocations.values()]
         assert len(set(invocation.task_id for invocation in invocations)) == 1
-        assert invocations[0].custom_image == custom_image_no_default["custom_image"]
+        assert invocations[0].custom_image == CUSTOM_IMAGE_NOT_DEFAULT["custom_image"]
 
         assert invocations[1].resources == invocations[0].resources
         assert invocations[1].custom_image == DEFAULT_IMAGE
@@ -347,8 +375,8 @@ class TestArifactFutureMethodsCalls:
         invocations = [*wf_model.task_invocations.values()]
         assert len(set(invocation.task_id for invocation in invocations)) == 1
         assert invocations[0].resources is not None
-        assert invocations[0].resources.dict() == resources_no_default
-        assert invocations[0].custom_image == custom_image_no_default["custom_image"]
+        assert invocations[0].resources.dict() == TASK_RESOURCES_EXPECTED
+        assert invocations[0].custom_image == CUSTOM_IMAGE_NOT_DEFAULT["custom_image"]
 
         assert invocations[1].resources is None
         assert invocations[1].custom_image is DEFAULT_IMAGE